Making RAMCloud compile with g++4.6
There were several points in the code (BackupFailureMonitor, BackupService,
MasterClient) that would not compile with a newer g++ than 4.4.4. They are
fixed now, the code compiles with g++4.4.4 as well as g++4.6.

+++ src/MasterClient.cc
@@ -414,7 +414,7 @@ MasterClient::multiRead(std::vector<ReadObject*> requests)
  *      May not be NULL. The new value of the object after incrementing.
  */
 void
-MasterClient::increment(uint32_t tableId, const char* key, uint16_t keyLength,
+MasterClient::increment(uint64_t tableId, const char* key, uint16_t keyLength,
                int64_t incrementValue, const RejectRules* rejectRules,
                uint64_t* version, int64_t* newValue)
 {
