+++ src/RamCloud.h
@@ -151,7 +151,7 @@ class RamCloud {
     void read(uint64_t tableId, const char* key, uint16_t keyLength,
               Buffer* value, const RejectRules* rejectRules = NULL,
               uint64_t* version = NULL);
-    void increment(uint32_t tableId, const char* key, uint16_t keyLength,
+    void increment(uint64_t tableId, const char* key, uint16_t keyLength,
               int64_t incrementValue, const RejectRules* rejectRules = NULL,
               uint64_t* version = NULL, int64_t* newValue = NULL);
     void multiRead(MasterClient::ReadObject* requests[], uint32_t numRequests);
